backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Google Gemini AI
google-generativeai==0.3.1

# Whisper for speech-to-text
openai-whisper==20231117
torch>=2.2.0
torchaudio>=2.2.0

# Audio/Video processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Data handling
pydantic==2.5.0
python-dotenv==1.0.0

# Database (optional, for future use)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Additional utilities
numpy==1.26.2
pandas==2.1.3

# Acoustic analysis
librosa==0.10.1
scipy==1.11.4
matplotlib==3.8.2

# Authentication
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF generation
reportlab==4.0.7